config: Added CommandPreProc (allows shell vars)
This commit adds a new optional Config field that allows developers
to intercept the list of strings typed by a user prior to looking up
a command and running the relevant code.

The main goal of this code is to allow variable replacement,
though it can be potentially used for other useful functionality.

Please refer to the code comments for more information on the
desired behavior of this code.